+++ core/bot.py
@@ -6,6 +6,7 @@ import json
 import logging
 import os
 import re
+import warnings
 from collections import Counter
 from contextlib import suppress
 from typing import Any, Dict, Iterable, List, Optional, Union
@@ -60,7 +61,7 @@ class ziBot(commands.Bot):
 
     # --- NOTE: Information about the bot
     author: str = getattr(config, "author", "ZiRO2264#9999")
-    version: str = "`3.2.9` - `overhaul`"
+    version: str = "`3.3.0` - `overhaul`"
     links: Dict[str, str] = getattr(
         config,
         "links",
@@ -74,6 +75,9 @@ class ziBot(commands.Bot):
     # ---
 
     def __init__(self) -> None:
+        # custom intents, required since dpy v1.5
+        intents = discord.Intents.all()
+
         super().__init__(
             command_prefix=_callablePrefix,
             description=(
@@ -81,9 +85,10 @@ class ziBot(commands.Bot):
                 "created by ZiRO2264, formerly called `ziBot`."
             ),
             case_insensitive=True,
-            intents=discord.Intents.all(),
+            intents=intents,
             heartbeat_timeout=150.0,
         )
+
         # make cogs case insensitive
         self._BotBase__cogs: commands.core._CaseInsensitiveDict = (
             commands.core._CaseInsensitiveDict()
@@ -98,7 +103,7 @@ class ziBot(commands.Bot):
 
         # Bot master(s)
         # self.master = (186713080841895936,)
-        self.master: tuple = (
+        self.owner_ids: tuple = (
             tuple()
             if not hasattr(config, "botMasters")
             else tuple([int(master) for master in config.botMasters])
@@ -197,16 +202,17 @@ class ziBot(commands.Bot):
     async def startUp(self) -> None:
         """Will run when the bot ready"""
         await self.wait_until_ready()
-        if not self.master:
+
+        if not self.owner_ids:
             # If self.master not set, warn the hoster
             self.logger.warning(
                 "No master is set, you may not able to use certain commands! (Unless you own the Bot Application)"
             )
 
-        # Add application owner into bot master list
+        # Add application owner into owner_ids list
         owner: discord.User = (await self.application_info()).owner
-        if owner and owner.id not in self.master:
-            self.master += (owner.id,)
+        if owner and owner.id not in self.owner_ids:
+            self.owner_ids += (owner.id,)
 
         # change bot's presence into guild live count
         self.changing_presence.start()
@@ -215,6 +221,11 @@ class ziBot(commands.Bot):
 
         if not hasattr(self, "uptime"):
             self.uptime: datetime.datetime = utcnow()
+
+    @property
+    def master(self):
+        warnings.warn("Bot.master is deprecated, use self.owner_ids instead")
+        return self.owner_ids
 
     async def getGuildConfigs(
         self, guildId: int, filters: Iterable = "*", table: str = "guildConfigs"
@@ -609,7 +620,7 @@ class ziBot(commands.Bot):
             message.author.bot
             or message.author.id in self.blacklist.users
             or (message.guild and message.guild.id in self.blacklist.guilds)
-        ) and message.author.id not in self.master:
+        ) and message.author.id not in self.owner_ids:
             return
 
         # if bot is mentioned without any other message, send prefix list
@@ -636,7 +647,7 @@ class ziBot(commands.Bot):
             message.author.bot
             or message.author.id in self.blacklist.users
             or (message.guild and message.guild.id in self.blacklist.guilds)
-        ) and message.author.id not in self.master:
+        ) and message.author.id not in self.owner_ids:
             return
 
         await self.process(message)