+++ exts/admin/admin.py
@@ -28,11 +28,14 @@ ROLE_TYPES = {
 
 
 class Admin(commands.Cog, CogMixin):
-    """Admin-only commands to configure the bot."""
+    """Collection of commands for admin/mods to configure the bot.
+
+    Some commands may require `Administrator` permission.
+    """
 
     icon = "\u2699"
 
-    def __init__(self, bot):
+    def __init__(self, bot) -> None:
         super().__init__(bot)
 
     @staticmethod
@@ -44,12 +47,7 @@ class Admin(commands.Cog, CogMixin):
         return await bot.setGuildConfig(guildId, roleType, roleId, "guildRoles")
 
     async def cog_check(self, ctx):
-        if not ctx.guild:
-            # Configuration only for Guild
-            return False
-
-        # Check if member can manage_channels or have mod role
-        return await checks.isMod(ctx)
+        return ctx.guild is not None
 
     async def handleGreetingConfig(self, ctx, arguments, type: str):
         """Handle welcome and farewell configuration."""
@@ -214,11 +212,12 @@ class Admin(commands.Cog, CogMixin):
                 "disable": "Disable modlog",
             },
             perms={
-                "bot": "Manage Channels",
+                "bot": "Manage Channels and View Audit Log",
                 "user": "Moderator Role or Manage Channels",
             },
         ),
     )
+    @commands.bot_has_guild_permissions(view_audit_log=True, manage_channels=True)
     @checks.mod_or_permissions(manage_channels=True)
     async def modlog(self, ctx, *, arguments):
         await self.handleLogConfig(ctx, arguments, "modlog")
@@ -435,6 +434,8 @@ class Admin(commands.Cog, CogMixin):
             },
         ),
     )
+    @commands.bot_has_guild_permissions(manage_roles=True)
+    @checks.is_admin()
     async def autorole(self, ctx, name: Union[discord.Role, str]):
         await ctx.try_invoke(
             self.roleCreate if isinstance(name, str) else self.roleSet,
@@ -451,6 +452,7 @@ class Admin(commands.Cog, CogMixin):
             },
         ),
     )
+    @checks.is_mod()
     async def announcement(self, ctx, channel: discord.TextChannel):
         await self.bot.setGuildConfig(
             ctx.guild.id, "announcementCh", channel.id, "guildChannels"