+++ exts/info/info.py
@@ -14,6 +14,7 @@ from aiohttp import InvalidURL, client_exceptions
 from discord.ext import commands
 
 from core import checks
+from core.context import Context
 from core.converter import MemberOrUser
 from core.embed import ZEmbed
 from core.mixin import CogMixin
@@ -21,20 +22,7 @@ from utils import pillow
 from utils.api.openweather import CityNotFound, OpenWeatherAPI
 from utils.format import formatDiscordDT, renderBar
 from utils.infoQuote import *  # noqa:  F403
-from utils.other import utcnow
-
-
-# TODO: Move this somewhere in `exts/utils/` folder
-async def authorOrReferenced(ctx):
-    if ref := ctx.replied_reference:
-        # Get referenced message author
-        # if user reply to a message while doing this command
-        return (
-            ref.cached_message.author
-            if ref.cached_message
-            else (await ctx.fetch_message(ref.message_id)).author
-        )
-    return ctx.author
+from utils.other import authorOrReferenced, utcnow
 
 
 class Info(commands.Cog, CogMixin):
@@ -53,32 +41,31 @@ class Info(commands.Cog, CogMixin):
         aliases=("av", "userpfp", "pfp"), brief="Get member's avatar image"
     )
     @commands.cooldown(1, 5, commands.BucketType.user)
-    async def avatar(self, ctx, _user: MemberOrUser = None):
-        user: Union[discord.Member, discord.User] = _user or await authorOrReferenced(
-            ctx
-        )
-        user.avatar.with_format("jpg").url
+    async def avatar(self, ctx: Context, user: MemberOrUser = None):
+        user = user or await authorOrReferenced(ctx)  # type: ignore
+
+        avatar: discord.Asset = user.display_avatar  # type: ignore
 
         # Links to avatar (with different formats)
         links = (
             "[`JPEG`]({})"
             " | [`PNG`]({})"
             " | [`WEBP`]({})".format(
-                user.avatar.with_format("jpg").url,
-                user.avatar.with_format("png").url,
-                user.avatar.with_format("webp").url,
+                avatar.with_format("jpg").url,
+                avatar.with_format("png").url,
+                avatar.with_format("webp").url,
             )
         )
-        if user.avatar.is_animated():
-            links += " | [`GIF`]({})".format(user.avatar.with_format("gif").url)
+        if avatar.is_animated():
+            links += " | [`GIF`]({})".format(avatar.with_format("gif").url)
 
         # Embed stuff
         e = ZEmbed.default(
             ctx,
-            title="{}'s Avatar".format(user.name),
+            title="{}'s Avatar".format(user.name),  # type: ignore
             description=links,
         )
-        e.set_image(url=user.avatar.with_size(1024).url)
+        e.set_image(url=avatar.with_size(1024).url)
         await ctx.try_reply(embed=e)
 
     @commands.command(
@@ -87,7 +74,7 @@ class Info(commands.Cog, CogMixin):
         extras=dict(example=("weather Palembang", "w London")),
     )
     @commands.cooldown(1, 5, commands.BucketType.user)
-    async def weather(self, ctx, *, city):
+    async def weather(self, ctx: Context, *, city):
         if not self.openweather.apiKey:
             return await ctx.error(
                 "OpenWeather's API Key is not set! Please contact the bot owner to solve this issue."
@@ -126,7 +113,7 @@ class Info(commands.Cog, CogMixin):
         extras=dict(example=("colour ffffff", "clr 0xffffff", "color #ffffff")),
     )
     @commands.cooldown(1, 5, commands.BucketType.user)
-    async def colour(self, ctx, value: str):
+    async def colour(self, ctx: Context, value: str):
         # Pre processing
         value = value.lstrip("#")[:6]
         value = value.ljust(6, "0")
@@ -177,7 +164,9 @@ class Info(commands.Cog, CogMixin):
         invoke_without_command=True,
     )
     @commands.cooldown(1, 5, commands.BucketType.user)
-    async def emoji(self, ctx, emoji: Union[discord.Emoji, discord.PartialEmoji, str]):
+    async def emoji(
+        self, ctx: Context, emoji: Union[discord.Emoji, discord.PartialEmoji, str]
+    ):
         # TODO: Add emoji list
         await ctx.try_invoke(self.emojiInfo, emoji)
 
@@ -195,9 +184,7 @@ class Info(commands.Cog, CogMixin):
         ),
     )
     @commands.cooldown(1, 5, commands.BucketType.user)
-    async def emojiInfo(
-        self, ctx, emoji: Union[discord.Emoji, discord.PartialEmoji, str]
-    ):
+    async def emojiInfo(self, ctx: Context, emoji: Union[discord.PartialEmoji, str]):
         try:
             e = ZEmbed.default(
                 ctx,
@@ -206,15 +193,16 @@ class Info(commands.Cog, CogMixin):
             )
             e.set_image(url=emoji.url)  # type: ignore
         except AttributeError:
-            # TODO: Doesn't work with :rock:, find a fix
             try:
+                strEm: str = str(emoji)
+                digit = f"{ord(strEm):x}"
                 e = ZEmbed.default(
                     ctx,
                     title=" - ".join(
                         (
-                            emoji,
-                            hex(ord(emoji)).replace("0x", r"\u"),
-                            unicodedata.name(emoji),
+                            strEm,
+                            f"\\u{digit}",
+                            unicodedata.name(strEm, "Name not found"),
                         )
                     ),
                     description="`Type: Unicode`",
@@ -237,11 +225,13 @@ class Info(commands.Cog, CogMixin):
     )
     @commands.guild_only()
     @checks.mod_or_permissions(manage_emojis=True)
-    async def emojiSteal(self, ctx, emoji: Union[discord.Emoji, discord.PartialEmoji]):
+    async def emojiSteal(
+        self, ctx: Context, emoji: Union[discord.Emoji, discord.PartialEmoji]
+    ):
         emojiByte = await emoji.read()
 
         try:
-            addedEmoji = await ctx.guild.create_custom_emoji(
+            addedEmoji = await ctx.guild.create_custom_emoji(  # type: ignore
                 name=emoji.name, image=emojiByte
             )
         except discord.Forbidden:
@@ -285,7 +275,7 @@ class Info(commands.Cog, CogMixin):
             except AttributeError:
                 # Probably a url?
                 try:
-                    async with self.bot.session.get(emoji) as req:
+                    async with self.bot.session.get(emoji) as req:  # type: ignore
                         emojiByte = await req.read()
                 except InvalidURL:
                     return await ctx.error(
@@ -331,7 +321,7 @@ class Info(commands.Cog, CogMixin):
         ),
     )
     @commands.cooldown(1, 5, commands.BucketType.user)
-    async def jisho(self, ctx, *, words):
+    async def jisho(self, ctx: Context, *, words):
         async with ctx.bot.session.get(
             "https://jisho.org/api/v1/search/words", params={"keyword": words}
         ) as req:
@@ -370,7 +360,7 @@ class Info(commands.Cog, CogMixin):
         brief="Show covid information on certain country",
     )
     @commands.cooldown(1, 5, commands.BucketType.user)
-    async def covid(self, ctx, *, country):
+    async def covid(self, ctx: Context, *, country):
         # TODO: Remove later
         if country.lower() in ("united state of america", "america"):
             country = "US"
@@ -400,9 +390,10 @@ class Info(commands.Cog, CogMixin):
         brief="Get user's information",
     )
     @commands.cooldown(1, 5, commands.BucketType.user)
-    async def userinfo(self, ctx, *, user: MemberOrUser = None):
-        if not user:
-            user = await authorOrReferenced(ctx)
+    async def userinfo(self, ctx: Context, *, _user: MemberOrUser = None):
+        user: Union[discord.User, discord.Member] = _user or await authorOrReferenced(
+            ctx
+        )  # type: ignore # MemberOrUser or authorOrReferenced will return user/member
 
         def status(x):
             return {
@@ -447,19 +438,22 @@ class Info(commands.Cog, CogMixin):
             }.get(x.type, "") + detail
 
         badges = [badge(x) for x in user.public_flags.all()]
-        if user == ctx.guild.owner:
+        if (guild := ctx.guild) and user == guild.owner:
             badges.append("<:owner:747802537402564758>")
 
         createdAt = user.created_at.replace(tzinfo=dt.timezone.utc)
         isUser = isinstance(user, discord.User)
         joinedAt = None
         if not isUser:
-            joinedAt = user.joined_at.replace(tzinfo=dt.timezone.utc)
+            # already checked by isUser
+            joinedAt = user.joined_at.replace(tzinfo=dt.timezone.utc)  # type: ignore
+
+        avatar = user.display_avatar
 
         e = (
             ZEmbed()
-            .set_author(name=user, icon_url=user.avatar.url)
-            .set_thumbnail(url=user.avatar.url)
+            .set_author(name=user, icon_url=avatar.url)
+            .set_thumbnail(url=avatar.url)
         )
 
         e.add_field(
@@ -488,9 +482,9 @@ class Info(commands.Cog, CogMixin):
                         formatDiscordDT(joinedAt, "F"), formatDiscordDT(joinedAt, "R")
                     )
                     + "**Role count**: ({}/{})\n".format(
-                        len(user.roles), len(user.guild.roles)
+                        len(user.roles), len(user.guild.roles)  # type: ignore
                     )
-                    + "**Top role**: {}".format(user.top_role.mention)
+                    + "**Top role**: {}".format(user.top_role.mention)  # type: ignore
                 )
             ),
             inline=False,
@@ -502,9 +496,9 @@ class Info(commands.Cog, CogMixin):
                 "N/A"
                 if isUser
                 else (
-                    "**Status**: {}\n".format(status(user.status))
+                    "**Status**: {}\n".format(status(user.status))  # type: ignore
                     + "**Activity**: {}".format(
-                        activity(user.activity) if user.activity else "None"
+                        activity(user.activity) if user.activity else "None"  # type: ignore
                     )
                 )
             ),
@@ -543,11 +537,10 @@ class Info(commands.Cog, CogMixin):
 
             status[str(m.status)][0] += 1
 
-        e = (
-            ZEmbed()
-            .set_author(name=guild, icon_url=guild.icon.url)
-            .set_thumbnail(url=guild.icon.url)
-        )
+        e = ZEmbed()
+
+        if icon := guild.icon:
+            e.set_author(name=guild, icon_url=icon.url).set_thumbnail(url=icon.url)
 
         e.add_field(
             name="General",
@@ -558,7 +551,7 @@ class Info(commands.Cog, CogMixin):
                     formatDiscordDT(createdAt, "F"), formatDiscordDT(createdAt, "R")
                 )
                 + "**Owner**: {0} / {0.mention}\n".format(guild.owner)
-                + "**Owner ID**: `{}`".format(guild.owner.id)
+                + "**Owner ID**: `{}`".format(guild.owner.id)  # type: ignore
             ),
             inline=False,
         )
@@ -611,12 +604,12 @@ class Info(commands.Cog, CogMixin):
     )
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.guild_only()
-    async def spotifyinfo(self, ctx, user: discord.Member = None):
-        user = user or ctx.author
+    async def spotifyinfo(self, ctx: Context, user: discord.Member = None):
+        user = user or await authorOrReferenced(ctx)  # type: ignore
 
         spotify: discord.Spotify = discord.utils.find(
-            lambda s: isinstance(s, discord.Spotify), user.activities
-        )
+            lambda s: isinstance(s, discord.Spotify), user.activities  # type: ignore
+        )  # discord.Spotify is ActivityTypes
         if not spotify:
             return await ctx.error(
                 "{} is not listening to Spotify!".format(user.mention)
@@ -641,7 +634,7 @@ class Info(commands.Cog, CogMixin):
         # Bar stuff
         barLength = 5 if user.is_on_mobile() else 17
         bar = renderBar(
-            (cur.seconds / dur.seconds) * 100,
+            int(cur.seconds / dur.seconds) * 100,
             fill="─",
             empty="─",
             point="⬤",
@@ -672,7 +665,9 @@ class Info(commands.Cog, CogMixin):
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.guild_only()
     async def permissions(
-        self, ctx, memberOrRole: Union[discord.Member, discord.Role, str] = None
+        self,
+        ctx: Context,
+        memberOrRole: Union[discord.Member, discord.User, discord.Role, str] = None,
     ):
         if isinstance(memberOrRole, str) or memberOrRole is None:
             memberOrRole = ctx.author
@@ -694,7 +689,7 @@ class Info(commands.Cog, CogMixin):
             ),
         )
         with suppress(AttributeError):
-            e.set_thumbnail(url=memberOrRole.avatar.url)  # type: ignore
+            e.set_thumbnail(url=memberOrRole.display_avatar.url)  # type: ignore
 
         await ctx.try_reply(embed=e)
 
@@ -703,7 +698,7 @@ class Info(commands.Cog, CogMixin):
         usage="(project name)",
     )
     @commands.cooldown(1, 5, commands.BucketType.user)
-    async def pypi(self, ctx, project: str):
+    async def pypi(self, ctx: Context, project: str):
         async with self.bot.session.get(f"https://pypi.org/pypi/{project}/json") as res:
             try:
                 res = await res.json()